requirements.txt
pip~=24.2
pillow~=11.0.0
pygame~=2.6.1
pybind11
